# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/firebase/test":{"get":{"operationId":"AppController_testFirebase","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/firebase/test-firestore":{"post":{"operationId":"AppController_testFirestore","parameters":[],"responses":{"201":{"description":""}},"tags":["App"]},"get":{"operationId":"AppController_getFirestoreTest","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/firebase/test-auth":{"get":{"operationId":"AppController_testAuth","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/playground":{"get":{"operationId":"PlaygroundController_getPlayground","parameters":[],"responses":{"200":{"description":""}},"tags":["Playground"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/loginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgetPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/verify-otp":{"post":{"operationId":"AuthController_verifyOtp","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/update-profile":{"post":{"operationId":"AuthController_updateUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/user/me":{"get":{"operationId":"UserController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]}},"/user/verify-email":{"get":{"operationId":"UserController_verifyEmail","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/salon/create":{"post":{"operationId":"SalonController_createSalon","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SalonCreateDto"}}}},"responses":{"201":{"description":"The salon has been successfully created."}},"security":[{"bearer":[]}],"summary":"Create a new salon","tags":["Salon"]}},"/salon/update/{id}":{"patch":{"operationId":"SalonController_updateSalon","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SalonUpdateDto"}}}},"responses":{"200":{"description":"The salon has been successfully updated."}},"security":[{"bearer":[]}],"summary":"Update an existing salon","tags":["Salon"]}}},"info":{"title":"Salon Store API","description":"The Salon Store API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserRegistrationDto":{"type":"object","properties":{}},"loginDto":{"type":"object","properties":{}},"ForgetPasswordDto":{"type":"object","properties":{}},"UserUpdateDto":{"type":"object","properties":{}},"SalonCreateDto":{"type":"object","properties":{}},"SalonUpdateDto":{"type":"object","properties":{}}}}}